fixed upload file on current year change the version

+++ src/Sie/RegularBundle/Controller/ConsolidationSieController.php
@@ -401,18 +401,28 @@ class ConsolidationSieController extends Controller {
 
                     
                 }else{
-                  if ((strcmp(preg_replace('/\s+/', '', $aFileInfoSie[1]), preg_replace('/\s+/', '', $this->session->get('currentyear')-1))) == 0) {
-                    if (
-                          (strcmp(preg_replace('/\s+/', '', $aFileInfoSie[10]), preg_replace('/\s+/', '', '1.2.8'))) === 0
-                            and ( strcmp(preg_replace('/\s+/', '', $aFileInfoSie[12]), preg_replace('/\s+/', '', 'SIGED5'))) === 0    
-                    ){
+                     if(
+                        (
+                          (strcmp(preg_replace('/\s+/', '', $aFileInfoSie[10]), preg_replace('/\s+/', '', '1.2.9.1'))) === 0 || 
+                          (strcmp(preg_replace('/\s+/', '', $aFileInfoSie[10]), preg_replace('/\s+/', '', '1.2.9.2'))) === 0 
+                        )
+                            and ( strcmp(preg_replace('/\s+/', '', $aFileInfoSie[12]), preg_replace('/\s+/', '', 'SIGED8'))) === 0
+                      ){
 
-                    }else{
-                      $session->getFlashBag()->add('warningcons', 'El archivo ' . $aDataExtractFileUE[1] . ' presenta versión incorrecta para subir el archivo ');
-                      system('rm -fr ' . $dirtmp);
-                      return $this->redirect($this->generateUrl('consolidation_sie_web'));
-                    }
-                  }
+                      }else{
+                          if ((strcmp(preg_replace('/\s+/', '', $aFileInfoSie[1]), preg_replace('/\s+/', '', $this->session->get('currentyear')-1))) == 0) {
+                            if (
+                                  (strcmp(preg_replace('/\s+/', '', $aFileInfoSie[10]), preg_replace('/\s+/', '', '1.2.8'))) === 0
+                                    and ( strcmp(preg_replace('/\s+/', '', $aFileInfoSie[12]), preg_replace('/\s+/', '', 'SIGED5'))) === 0    
+                            ){
+
+                            }else{
+                              $session->getFlashBag()->add('warningcons', 'El archivo ' . $aDataExtractFileUE[1] . ' presenta versión incorrecta para subir el archivo ');
+                              system('rm -fr ' . $dirtmp);
+                              return $this->redirect($this->generateUrl('consolidation_sie_web'));
+                            }
+                          }
+                      }
                }
                 //die;
                 //validate the correct sie send with the content of file
